Rotation matrix utils / class
Same case as the push with the small quaternion library I pushed just now.

I don't have access to the BMI270 as of writing this, so you'll notice a lot of my code is "overly-explicit" -- this is highly intentional. 
Same sentiment behind this as with the quaternion push -- so make sure you read that.

Godspeed.